+++ src/PropertyPortfolioManager.Models/Model/Property/UnitTypeModel.cs
@@ -2,7 +2,7 @@
 {
     public class UnitTypeModel
     {
-        public Int16 Id { get; set; }
+        public int Id { get; set; }
 
         public string Type { get; set; } = string.Empty;
     }
